% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colrdata}
\alias{colrdata}
\title{Dummy data for demonstration of the capacities of 'rcol'}
\format{A data frame with 3 rows and 5 variables:
\describe{
  \item{01-31-1955}{dummy}
  \item{02-15-1980}{dummy}
  \item{06/02/1999}{dummy}
  \item{02 14 01}{dummy}
  \item{03 03-2016}{dummy}...
}}
\source{
my thumb
}
\usage{
colrdata
}
\description{
A dataset containing with columns with malformed dates.
}
\keyword{datasets}

